eval_law: pre-calculate num_lets.
Calculuate the number of let-bindings within a law at *law construction*
time, and then just use that during law evaluation.

In addition to counting the lets, this also counts the number of calls.

With this information, a later routine can preallocate the entire graph
produced by law substitution with a single big alloc.

A nice side-effect of that, will be that the code that walks the body
to fill the graph will no longer need to do futher allocs along the way,
so it doesn't need to be gc-safe.