Use only one alloc per law-evaluation.
Since we now pre-compute the number of let+call nodes needed for each law,
we can just do one allocation that allocates the entire graph.

Since we can do the whole allocation *before* graph construction, the
actual graph-construction doesn't need to be gc safe, which is much
simpler and faster.