Just one memory region.
Sometimes with the old approach, we would get segfaults because our
MAP_FIXED regions clobbered something that the stdlib placed randomly.
This fixes that by letting the OS choose the location of the loom.

Also, this eliminates the separate mmap() region used for the stack,
by just putting the stack directly in front of the heap on the loom.
Since the stack grows downwards, we will still get segfaults (not heap
corruption) on stack overflow.

Finally, this makes the heap executable (insecure, but simple is better
for now) and eliminates the jitspace heap.  This patch stubs out the
jitspace allocator.  That should be replacled with logic that just writes
code directly into the Law's allocation box on the heap.

NOTE: if we use this approach, is it possible to use relative addresses to
refer to constant values?  I don't know how that works, but if something
like that is possible, it could avoid the whole issue with needing to push
constants to the stack.  (e.g. after gc() the code would still be in
the same place relative to the constants table)