reorg stack for laws with lets.
This re-arranges the stack so that var-reference i is always at st[i].

This is zero work in the case of laws without lets (the common case),
but requires sliding the arguments in the case of laws *with* lets.

Contrary to the last commit message, this is slightly faster.

1.  One less branch per application per law.

2.  Laws with lets must do a small copy.

It seems that #1 pays for #2, but just barely.

The performance difference is insignificant, but this version seems
less complicated.